+++ src/main/java/nextstep/exception/newsectionexception/NewSectionException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.exception.newsectionexception;
 
 public class NewSectionException extends RuntimeException {
