+++ src/main/java/nextstep/exception/notfoundexception/NotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.exception.notfoundexception;
 
 public class NotFoundException extends RuntimeException {
