+++ src/main/java/nextstep/exception/notfoundexception/SectionNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.exception.notfoundexception;
 
 public class SectionNotFoundException extends NotFoundException {
