+++ src/main/java/nextstep/exception/notfoundexception/StationNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.exception.notfoundexception;
 
 public class StationNotFoundException extends NotFoundException {
