+++ src/main/java/nextstep/exception/pathexception/PathException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.exception.pathexception;
 
 public class PathException extends RuntimeException {
