+++ src/main/java/nextstep/exception/removesectionexception/NotExistedRemoveSectionException.java
@@ -1,4 +1,6 @@
-package nextstep.subway.exception;
+package nextstep.exception.removesectionexception;
+
+import nextstep.exception.removesectionexception.RemoveSectionException;
 
 public class NotExistedRemoveSectionException extends RemoveSectionException {
