+++ src/main/java/nextstep/subway/applicaion/StationService.java
@@ -1,10 +1,10 @@
 package nextstep.subway.applicaion;
 
+import nextstep.exception.notfoundexception.StationNotFoundException;
 import nextstep.subway.applicaion.dto.StationRequest;
 import nextstep.subway.applicaion.dto.StationResponse;
 import nextstep.subway.domain.Station;
 import nextstep.subway.domain.StationRepository;
-import nextstep.subway.exception.StationNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
